WIP Setup upgrade for development and packaging.
This was done in August 2023.
I tried improving the general setup for the IDE, TS and GLSL linting and being proper with having UI Renderer as a separate package.
Alas, it took me a long time and I was unhappy with the workarounds for ESLint's new flat config. It was still WIP and I eventually focused on something else.
I'll now commit this as-was to have it separated from new work on updating the setup. (flat-compat should go!)

From my Aug '23 notes:
- Has latest TS and ESLint with flat config.
- Unfortunately too early, so flatcompat. but it's still better than the issues I had with exports, conflicting and TS from vue. Watchtower is similar: one of the rulesets is disabled, but not as terrible because it doesn't have the nested docs package with Vue.
- Actual rules to enable still need investigation. Even 'semi' :)