Continuation of WIP Setup upgrade for development and packaging
Improvements to the configuration, but the projects isn't running due to an error in shader code.
Compiles with tsc and lints with ESLint (with errors).

Improved TS configuration.
Reworked ESLint setup in a similar way to uirenderer-canvas, but did not double check the resulting configuration.

Added missing dev dependencies that were causing import errors in configuration files.
- typescript-eslint is needed to get those two things working together.
- @eslint/js has the recommended JS linting configuration to use as base for the rest.

Updated dependencies.